change all react context to zustand
