+++ components/Maps/MapsControl.tsx
@@ -1,9 +1,8 @@
-import { useContext } from "react";
 import { ScaleControl, ZoomControl } from "react-leaflet";
-import { MapSettingsContext } from "@context/MapSettingsContext";
+import { useMapSettingsStore } from "@store/mapSettingsStore";
 
 export const MapControl = () => {
-  const { showControl } = useContext(MapSettingsContext)
+  const showControl = useMapSettingsStore(state => state.showControl)
 
   if (!showControl) return null;
