+++ components/Sidebar/Options/BaseMapsOptions.tsx
@@ -1,11 +1,11 @@
-import { useContext } from "react";
 import { Select } from "@mantine/core";
 import { ListItem } from "../Common/ListItem";
 import { baseMaps } from "@config/baseMaps";
-import { BaseMapContext } from "@context/BaseMapContext";
+import { useBaseMapStore } from "@store/baseMapStore";
 
 export function BaseMapsOptions() {
-  const { baseMap, setBaseMap } = useContext(BaseMapContext)
+  const baseMap = useBaseMapStore(state => state.baseMap)
+  const setBaseMap = useBaseMapStore(state => state.setBaseMap)
 
   const selectBaseMap = (value: string) => {
     const selectedBaseMap = baseMaps.find((item) => item.name === value) || null
