+++ components/Sidebar/Options/MapsControlOptions.tsx
@@ -1,10 +1,10 @@
-import { useContext } from "react";
-import { MapSettingsContext } from "@context/MapSettingsContext";
 import { ListItem } from "@components/Sidebar/Common/ListItem";
 import { Switcher } from "@components/Sidebar/Common/Switcher";
+import { useMapSettingsStore } from "@store/mapSettingsStore";
 
 export function MapsControlOptions() {
-  const { showControl, toggleControl } = useContext(MapSettingsContext)
+  const showControl = useMapSettingsStore(state => state.showControl)
+  const toggleControl = useMapSettingsStore(state => state.toggleControl)
   return (
     <ListItem title="Show Maps Control" description="Hide or show maps control, e.g zoom" >
       <Switcher checked={showControl} onChange={() => toggleControl()} />
