--- lib/provider/MapProvider.tsx
@@ -1,10 +1,0 @@
-import { BaseMapProvider } from '@context/BaseMapContext';
-import { MapSettingsProvider } from '@context/MapSettingsContext';
-
-export const MapProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => (
-  <MapSettingsProvider>
-    <BaseMapProvider>
-      {children}
-    </BaseMapProvider>
-  </MapSettingsProvider>
-)
